requirements.txt
jsonschema==2.6.0
enum34==1.1.6
lxml==3.5.0
mysql-connector-python-rf==2.0.4
namedlist==1.7
py==1.4.31
PyHamcrest==1.9.0
# pytest==2.9.2
pytest==3.0.3
pytest-xdist==1.15.0
pytest-timeout==1.0.0
https://github.com/allure-framework/allure-pytest/archive/master.zip

requests==2.11.1
rfc3987==1.3.4
selenium==2.53.2
six==1.10.0
wheel==0.26.0
fake-factory==0.5.2
configparser==3.5.0b2
beautifulsoup4==4.4.1
pyvirtualdisplay==0.2
Pillow==3.2.0
deepdiff==3.1.0
pre-commit
contextlib2==0.5.3
pytz==2016.4
retrying==1.3.3
python-dateutil==2.5.3
phpserialize==1.3
arrow==0.10.0
fuzzing==0.3.2

